переход с кошелька в форму регистрации + стили
