+++ shared/pages/CreateWallet/CreateWallet.scss
@@ -4,12 +4,6 @@
   margin: 0 auto;
 }
 
-.notYet {
-  text-align: center;
-  padding-top: 5px;
-  color: #6144e5;
-}
-
 .wrapper h2 {
   width: 400px;
   margin-bottom: 60px;
@@ -20,15 +14,9 @@
   margin: 0 16px;
 }
 
-.inputWrapper,
 .currencyChooserWrapper,
 .inLine {
   display: flex;
-}
-
-.inputWrapper {
-  justify-content: center;
-  margin: 50px 0;
 }
 
 .notMain {
@@ -132,10 +120,8 @@
 
 .currencyChooserWrapper {
   justify-content: space-between;
-}
-
-.currencyChooserWrapper {
   padding: 30px;
+
 }
 
 .card {
@@ -181,9 +167,27 @@
   width: 140px;
 }
 
+@media screen and (max-width: 750px){
+  .secureSize {
+    width: 100% !important;
+    height: 95px !important;
+
+  }
+  .currencyChooserWrapperSecond{
+    overflow-y: auto;
+    display: block;
+    padding: 0;
+  }
+
+  .currencyChooserWrapperSecond{
+    height: 225px;
+    margin: 10px 0;
+  }
+}
+
 @media screen and (max-width: 700px){
   .card {
-    width: 85px
+    width: 85px;
   }
 }
 
@@ -196,10 +200,12 @@
 
 @media screen and (max-width: 480px) {
   .continue {
-    position: absolute;
+    position: fixed;
+    z-index: 999999;
     right: 21px;
-    width: 89%;
-    top: 92vh;
+    width: 100%;
+    top: 95vh;
+    left: 0;
   }
 
   .secondStepInput {
@@ -238,7 +244,17 @@
     margin: 0;
   }
 
-  .third {
+  .currencyChooserWrapper {
+    overflow-y: auto;
+    display: block;
+    padding: 0;
+  }
+
+  .currencyChooserWrapper {
+    height: 50%;
+  }
+
+  .second {
     height: 59px;
     margin-bottom: 0px;
   }
@@ -266,7 +282,7 @@
   }
 
   .steps {
-    width: 150px
+    width: 100px
   }
 
   .logo {
@@ -274,7 +290,7 @@
   }
 
   .listGroup {
-    margin: auto 10px;
+    margin: auto;
   }
 
   .secureSize,
@@ -308,22 +324,6 @@
     height: 65px;
   }
 
-  .inputWrapper {
-    display: block;
-    margin: 0;
-  }
-
-  .currencyChooserWrapper{
-    overflow-y: auto;
-    display: block;
-    padding: 0;
-  }
-
-  .inputWrapper,
-  .currencyChooserWrapper{
-    height: 265px;
-  }
-
   .thirdPageIcons {
     background: #f0eefd;
   }
@@ -345,9 +345,7 @@
   }
 
   .secureChooserWrapper,
-  .inputWrapper,
   .currencyChooserWrapper {
-    height: 50%;
     overflow-y: auto;
   }
 
@@ -364,7 +362,8 @@
 
 @media screen and (max-width: 320px){
   .continue {
-    top: 91vh;
-    width: 87%;
+    top: 92vh;
+    width: 100%;
+    left: 0;
   }
 }
